Microsoft Power Bi
Then using Microsoft Power BI which is a  is a business analytics service that helps adeventureworkslt teams to  visualize and analyze data to find insights. It can connect to different data sources, transform and clean the data, and create charts, graphs, maps, and scatter plots to help identify patterns, trends, and anomalies.